refactor: Rename `lib/plugins/aws/utils/resolveLambdaTarget.js`
