+++ lib/plugins/aws/package/compile/events/alb/lib/targetGroups.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const resolveLambdaTarget = require('../../../../../utils/resolveLambdaTarget');
+const resolveLambdaTarget = require('../../../../../utils/resolve-lambda-target');
 
 const healthCheckDefaults = {
   HealthCheckEnabled: false,
