+++ lib/plugins/aws/package/compile/events/apiGateway/lib/validate.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const awsArnRegExs = require('../../../../../utils/arn-regular-expressions');
-const resolveLambdaTarget = require('../../../../../utils/resolveLambdaTarget');
+const resolveLambdaTarget = require('../../../../../utils/resolve-lambda-target');
 const ServerlessError = require('../../../../../../../serverless-error');
 const { log } = require('@serverless/utils/log');
