+++ lib/plugins/aws/package/compile/events/httpApi.js
@@ -6,7 +6,7 @@ const memoizee = require('memoizee');
 const memoizeeMethods = require('memoizee/methods');
 const { log } = require('@serverless/utils/log');
 const ServerlessError = require('../../../../../serverless-error');
-const resolveLambdaTarget = require('../../../utils/resolveLambdaTarget');
+const resolveLambdaTarget = require('../../../utils/resolve-lambda-target');
 
 const allowedMethods = new Set(['ANY', 'GET', 'POST', 'PUT', 'PATCH', 'OPTIONS', 'HEAD', 'DELETE']);
 const methodPattern = new RegExp(`^(?:\\*|${Array.from(allowedMethods).join('|')})$`, 'i');
