+++ lib/plugins/aws/package/compile/events/stream.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const resolveLambdaTarget = require('../../../utils/resolveLambdaTarget');
+const resolveLambdaTarget = require('../../../utils/resolve-lambda-target');
 
 class AwsCompileStreamEvents {
   constructor(serverless) {
