+++ lib/plugins/aws/utils/resolve-lambda-target.js
